move filesManagers to package filesManagers, refactor in both classes
