+++ src/main/java/com/jarq/system/managers/filesManagers/IContentReader.java
@@ -1,4 +1,4 @@
-package com.jarq.system.managers;
+package com.jarq.system.managers.filesManagers;
 
 import java.io.IOException;
