+++ src/main/java/com/jarq/system/managers/filesManagers/RepositoryFileReader.java
@@ -1,4 +1,4 @@
-package com.jarq.system.managers;
+package com.jarq.system.managers.filesManagers;
 
 import java.io.*;
 import java.nio.charset.Charset;
@@ -6,23 +6,21 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-public class RepositoryFileReader implements IContentReader<String> {
-
-    private final Charset ENCODING;
+public class RepositoryFileReader extends RepositoryFile implements IContentReader<String> {
 
     public static IContentReader<String> getInstance(Charset charset) {
         return new RepositoryFileReader(charset);
     }
 
     private RepositoryFileReader(Charset charset) {
-        this.ENCODING = charset;
+        super(charset);
     }
 
     @Override
     public String readContent(String filePath) throws IOException {
         Path path = Paths.get(filePath);
         StringBuilder stringBuilder = new StringBuilder();
-        try (BufferedReader reader = Files.newBufferedReader(path, ENCODING)) {
+        try (BufferedReader reader = Files.newBufferedReader(path, getEncoding())) {
             String line;
             while ((line = reader.readLine()) != null) {
                 stringBuilder.append(line);
