+++ src/main/java/com/jarq/system/managers/filesManagers/RepositoryFileWriter.java
@@ -1,4 +1,4 @@
-package com.jarq.system.managers;
+package com.jarq.system.managers.filesManagers;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
@@ -7,22 +7,20 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-public class RepositoryFileWriter implements IContentWriter<String> {
-
-    private final Charset ENCODING;
+public class RepositoryFileWriter extends RepositoryFile implements IContentWriter<String> {
 
     public static IContentWriter<String> getInstance(Charset charset) {
         return new RepositoryFileWriter(charset);
     }
 
     private RepositoryFileWriter(Charset charset) {
-        this.ENCODING = charset;
+        super(charset);
     }
 
     @Override
     public boolean writeContent(String filePath, String content) throws IOException {
         Path path = Paths.get(filePath);
-        try (BufferedWriter writer = Files.newBufferedWriter(path, ENCODING) ) {
+        try (BufferedWriter writer = Files.newBufferedWriter(path, getEncoding()) ) {
                 writer.write(content);
         }
         return true;
